Drop special case to avoid replacing ArtMethod references on stack
When replacing strong roots there was a special case for avoiding
obsolete methods that are still on the stack. These methods declaring
class shouldn't be changed. This logic was brittle and we might miss
other places where we could see obsolete ArtMethods. After this CL:
https://android-review.googlesource.com/c/platform/art/+/2134336
we update the obsolete method's declaring class after replacing
references. So this special logic is no longer necessary.

Test: art/test.py
Change-Id: Id7f678c4916acdf85fc46ac3b7f5948f4dc98631
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>